exposes the login_url method to allow people to create their own auth button/link fixes https://github.com/dwyl/hapi-auth-github/issues/7
