+++ lib/index.js
@@ -1,7 +1,7 @@
 var OPTIONS;      // global object set when plugin loads
 var request      = require('./http_request');
 var querystring  = require('querystring'); // nodejs.org/api/querystring.html
-// var redis_client = require('redis-connection')();
+var crypto      = require('crypto');    // http://nodejs.org/api/crypto.html
 
 /**
  * make_post_data simply extracts the code and state (hash) from the request
@@ -92,3 +92,17 @@ exports.register = function githubauth (server, options, next) {
 exports.register.attributes = {
     pkg: require('../package.json')
 };
+
+module.exports.login_url = function() {
+  var params = {
+    client_id : process.env.GITHUB_CLIENT_ID,
+    redirect_uri : process.env.BASE_URL + OPTIONS.REDIRECT_URL,
+    scope : 'repo',
+    state: crypto.createHash('sha256').update(Math.random().toString()).digest('hex')
+  }
+  console.log(params);
+  var qs = querystring.stringify(params);
+  console.log(qs);
+  var url = 'https://github.com/login/oauth/authorize' + '?' + qs;
+  return url;
+}
